expose helpers as library
